MozReview-Commit-ID: IHPBV4z9vPi

+++ xpcom/threads/HangMonitor.cpp
@@ -163,7 +163,7 @@ GetChromeHangReport(Telemetry::ProcessedStack& aStack,
   }
 
   if (!suspended) {
-    if (ret != -1) {
+    if (ret != (DWORD)-1) {
       MOZ_ALWAYS_TRUE(::ResumeThread(winMainThreadHandle) != DWORD(-1));
     }
     return;
@@ -172,7 +172,7 @@ GetChromeHangReport(Telemetry::ProcessedStack& aStack,
   MozStackWalkThread(ChromeStackWalker, /* skipFrames */ 0, /* maxFrames */ 0,
                      &rawStack, winMainThreadHandle, nullptr);
   ret = ::ResumeThread(winMainThreadHandle);
-  if (ret == -1) {
+  if (ret == (DWORD)-1) {
     return;
   }
   aStack = Telemetry::GetStackAndModules(rawStack);
